+++ packages/core/src/interpreter.ts
@@ -533,6 +533,11 @@ export class Interpreter<
     this.doneListeners.clear();
     this.errorListeners.clear();
 
+    if (!this.initialized) {
+      // Interpreter already stopped; do nothing
+      return this;
+    }
+
     this.state.configuration.forEach((stateNode) => {
       for (const action of stateNode.definition.exit) {
         this.exec(action, this.state);