+++ packages/xstate-dev/src/extension/devtools.js
@@ -1,1 +1,0 @@
-chrome.devtools.panels.create('XState', null, 'panel/build/index.html');