+++ packages/xstate-fsm/src/index.ts
@@ -150,13 +150,13 @@ export function createMachine<
     },
     transition: (
       state: string | StateMachine.State<TContext, TEvent, TState>,
-      event: string | (Record<string, any> & { type: string })
+      event: TEvent | TEvent['type']
     ): StateMachine.State<TContext, TEvent, TState> => {
       const { value, context } =
         typeof state === 'string'
           ? { value: state, context: fsmConfig.context! }
           : state;
-      const eventObject = toEventObject(event);
+      const eventObject = toEventObject<TEvent>(event);
       const stateConfig = fsmConfig.states[value];
 
       if (!IS_PRODUCTION) {
@@ -170,26 +170,32 @@ export function createMachine<
       }
 
       if (stateConfig.on) {
-        const transitions = toArray(stateConfig.on[eventObject.type]);
+        const transitions: Array<
+          StateMachine.Transition<TContext, TEvent>
+        > = toArray(stateConfig.on[eventObject.type]);
 
         for (const transition of transitions) {
           if (transition === undefined) {
             return createUnchangedState(value, context);
           }
 
-          const { target = value, actions = [], cond = () => true } =
+          const { target, actions = [], cond = () => true } =
             typeof transition === 'string'
               ? { target: transition }
               : transition;
 
+          const isTargetless = target === undefined;
+
           if (cond(context, eventObject)) {
-            const nextStateConfig = fsmConfig.states[target];
-            const allActions = ([] as any[])
-              .concat(stateConfig.exit, actions, nextStateConfig.entry)
-              .filter((a) => a)
-              .map<StateMachine.ActionObject<TContext, TEvent>>((action) =>
-                toActionObject(action, (machine as any)._options.actions)
-              );
+            const nextStateConfig = fsmConfig.states[target ?? value];
+            const allActions = (isTargetless
+              ? toArray(actions)
+              : ([] as any[])
+                  .concat(stateConfig.exit, actions, nextStateConfig.entry)
+                  .filter((a) => a)
+            ).map<StateMachine.ActionObject<TContext, TEvent>>((action) =>
+              toActionObject(action, (machine as any)._options.actions)
+            );
 
             const [nonAssignActions, nextContext, assigned] = handleActions(
               allActions,
@@ -197,13 +203,15 @@ export function createMachine<
               eventObject
             );
 
+            const resolvedTarget = target ?? value;
+
             return {
-              value: target,
+              value: resolvedTarget,
               context: nextContext,
               actions: nonAssignActions,
               changed:
                 target !== value || nonAssignActions.length > 0 || assigned,
-              matches: createMatcher(target)
+              matches: createMatcher(resolvedTarget)
             };
           }
         }